feat: extend existing patch-diff-links feature
