+++ source/features/patch-diff-links.tsx
@@ -1,9 +1,38 @@
 import React from 'dom-chef';
 import * as pageDetect from 'github-url-detection';
 
+import DiffIcon from 'octicons-plain-react/Diff';
+
 import features from '../feature-manager.js';
 import {getCleanPathname} from '../github-helpers/index.js';
 import observe from '../helpers/selector-observer.js';
+
+// PR URL resolution
+function getPrUrl(extension: 'patch' | 'diff'): string {
+	const pathname = getCleanPathname();
+	const [owner, repo, , id] = pathname.split('/');
+	return `/${owner}/${repo}/pull/${id}.${extension}`;
+}
+
+function updatePrUrl(
+	event: React.FocusEvent<HTMLAnchorElement> | React.MouseEvent<HTMLAnchorElement>,
+): void {
+	const link = event.currentTarget;
+	link.href = getPrUrl(link.textContent as 'patch' | 'diff');
+}
+
+function createPrLink(type: 'patch' | 'diff'): JSX.Element {
+	return (
+		<a
+			href={getPrUrl(type)}
+			className="sha color-fg-default"
+			onMouseEnter={updatePrUrl}
+			onFocus={updatePrUrl}
+		>
+			{type}
+		</a>
+	);
+}
 
 function getCommitUrl(extension: 'patch' | 'diff'): string {
 	// The replacement avoids a redirection isPRCommit
@@ -45,16 +74,37 @@ async function addPatchDiffLinks(commitMeta: HTMLElement): Promise<void> {
 	);
 }
 
+async function addPrPatchDiffLinks(prHeader: HTMLElement): Promise<void> {
+	prHeader.append(
+		<li className="mx-0 tmp-py-3 tmp-px-4 width-full rounded-0">
+			<div className="d-flex flex-items-center text-bold">
+				<DiffIcon className="mr-2" />
+				<div className="sha-block" data-turbo="false">
+					{createPrLink('patch')}
+					{' '}
+					{createPrLink('diff')}
+				</div>
+			</div>
+		</li>,
+	);
+}
+
 async function init(signal: AbortSignal): Promise<void> {
 	observe([
 		'.commit-meta > :is(span, div):last-child', // `isPRCommit` + old `isSingleCommit`
 		'[class*="commit-header-actions"] + div pre',
 	], addPatchDiffLinks, {signal});
+
+	observe([
+		'.react-overview-code-button-action-list > ul', // PR-specific selector
+		'#local-panel',
+	], addPrPatchDiffLinks, {signal});
 }
 
 void features.add(import.meta.url, {
 	include: [
 		pageDetect.isCommit,
+		pageDetect.isPR,
 	],
 	exclude: [
 		pageDetect.isPRCommit404,
@@ -68,5 +118,6 @@ Test URLs:
 
 - Commit:https://github.com/refined-github/refined-github/commit/132272786fdc058193e089d8c06f2a158844e101
 - PR Commit: https://github.com/refined-github/refined-github/pull/7751/commits/07ddf838c211075701e9a681ab061a158b05ee79
+- PR: https://github.com/refined-github/refined-github/pull/7751
 
 */
