+++ source/refined-github.ts
@@ -220,4 +220,3 @@ import './features/linkify-text.js';
 import './features/actions-run-removal.js';
 import './features/new-milestone-button.js';
 import './features/no-self-reference.js';
-import './features/pr-patch-diff-links.js';
